WeakHashMap으로 읽는 파일들을 캐싱
