+++ app/src/main/java/net/jspiner/epub_viewer/ui/search/finder/PagePageFinder.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import io.reactivex.Single
 import net.jspiner.epub_viewer.dto.Epub
 import net.jspiner.epub_viewer.dto.PageInfo
+import net.jspiner.epub_viewer.util.readFile
 import net.jspiner.epubstream.dto.ItemRef
 
 class PagePageFinder(context: Context, epub: Epub, pageInfo: PageInfo) : PageFinder(context, epub, pageInfo) {
